% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capture}
\alias{capture}
\title{Fish Capture Data}
\format{
A tbl data frame:
\describe{
\item{Capture}{The unique fish code (fctr).}
\item{DateTimeCapture}{The date and time of capture (time).}
\item{SectionCapture}{The section code (fint).}
\item{Species}{The fish species 'Bull Trout', 'Lake Trout' or
'Rainbow Trout' (fctr).}
\item{Length}{The fork length in mm (int).}
\item{Weight}{The wet mass in kg (dbl).}
\item{Reward1}{The reward value of the first T-Bar tag
in Canadian dollars (int).}
\item{Reward2}{The reward value of the second T-Bar tag if present
in Canadian dollars (int).}
\item{DateTimeTagExpire}{The acoustic tag expiration date and time (time).}
}
}
\usage{
capture
}
\description{
Fish Capture Data
}
\keyword{datasets}
