% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deployment}
\alias{deployment}
\title{Receiver Deployment Data}
\format{
A tbl data frame:
\describe{
\item{Station}{The station name (fctr).}
\item{Receiver}{The receiver code (fctr).}
\item{DateTimeReceiverIn}{The receiver deployment date and time (time).}
\item{DateTimeReceiverOut}{The receiver retrieval date and time (time).}
}
}
\usage{
deployment
}
\description{
A data frame of receiver deployments by station and date times.
}
\keyword{datasets}
