% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{depth}
\alias{depth}
\title{Acoustic Depth Dat}
\format{A tbl data frame:
\describe{
  \item{DateTimeDepth}{The depth date and time (time).}
  \item{Capture}{The fish capture number (int).}
  \item{Receiver}{The receiver number (int).}
  \item{Depth}{The depth (dbl).}
}}
\usage{
depth
}
\description{
Depth detection data by fish (capture) and receiver.
}
\keyword{datasets}

