% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{detection}
\alias{detection}
\title{Acoustic Detection Data}
\format{
A tbl data frame:
\describe{
\item{DateTimeDetection}{The detection date and hour (time).}
\item{Capture}{The fish code (fctr).}
\item{Receiver}{The receiver code (fctr).}
\item{Detections}{The number of detections in the hour (int).}
}
}
\usage{
detection
}
\description{
Hourly acoustic detection data by fish (capture) and receiver.
}
\keyword{datasets}
