% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{klexdatr}
\alias{klexdatr}
\alias{klexdatr-package}
\title{Kootenay Lake Exploitation Study Data}
\description{
For a description of the available data sets type \code{data(package = "klexdatr")}.
}

