% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{qlexdatr}
\alias{qlexdatr}
\alias{qlexdatr-package}
\title{Quesnel Lake Exploitation Study Data}
\description{
For a description of the available data sets type \code{data(package = "qlexdatr")}.
}

