% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recapture}
\alias{recapture}
\title{Fish Recapture Data}
\format{
A tbl data frame:
\describe{
\item{DateTimeRecapture}{The reported date of recapture (time).}
\item{Capture}{The fish code (fctr).}
\item{SectionRecapture}{The section code (fctr).}
\item{TBarTag1}{The first T-Bar Tag was reported (lgl).}
\item{TBarTag2}{A second T-Bar Tag was reported (lgl).}
\item{TagsRemoved}{The T-Bar tags were removed from the fish (lgl).}
\item{Released}{The angler reportedly released the fish (lgl).}
\item{Public}{The angler was a member of the public as opposed
the study team (lgl).}
}
}
\usage{
recapture
}
\description{
A tbl data frame of fish recaptures. As the time of recapture
was not reported it is assumed to be 12:00:00.
}
\keyword{datasets}
