% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{section}
\alias{section}
\title{Section Data}
\format{
A SpatialPolygonsDataFrame with the data frame:
\describe{
\item{Section}{The unique section code (fctr).}
\item{Habitat}{The habitat type 'Lentic' or 'Lotic' (fctr).}
\item{Bounded}{The polygon represents the full area (lgl).}
\item{geometry}{The section polygon (MULTIPOLYGON (m)).}
}
}
\usage{
section
}
\description{
Section Spatial Polygon Data
}
\details{
Polygons of sections of the waterbodies.
}
\keyword{datasets}
