% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{station}
\alias{station}
\title{Station Data}
\format{
A tbl data frame:
\describe{
\item{Station}{The unique station name (fctr).}
\item{Section}{The section code (fctr).}
\item{geometry}{The station point (POINT (m)).}
}
}
\usage{
station
}
\description{
A tbl data frame of detection stations.
}
\keyword{datasets}
