% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klexdatr-package.R
\docType{package}
\name{klexdatr-package}
\alias{klexdatr}
\alias{klexdatr-package}
\title{klexdatr: Kootenay Lake Exploitation Study Data}
\description{
An R package of data from the Kootenay Lake Large Trout Exploitation Study.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/klexdatr}
  \item Report bugs at \url{https://github.com/poissonconsulting/klexdatr/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} [data contributor]

Authors:
\itemize{
  \item Greg Andrusak [data contributor]
  \item Gary Pavan [data contributor]
  \item Sarah Stephenson [data contributor]
  \item Matt Neufeld [data contributor]
  \item Jeff Burrows [data contributor]
  \item Kerry Reed [data contributor]
  \item Robyn Irvine [data contributor]
  \item Harvey Andrusak [data contributor]
}

}
\keyword{internal}
