% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klexdatr-package.R
\docType{package}
\name{klexdatr-package}
\alias{klexdatr}
\alias{klexdatr-package}
\title{klexdatr: Kootenay Lake Exploitation Study Data}
\description{
Six relational 'tibbles' from the Kootenay Lake Large Trout Exploitation study.
  The study which ran from 2008 to 2014 caught, tagged and released large Rainbow Trout and Bull Trout
  in Kootenay Lake by boat angling. 
  The fish were tagged with internal acoustic tags and/or high reward external tags
  and subsequently detected by an acoustic receiver array as well as reported by anglers.
  The data are analysed by Thorley and Andrusak (1994) <doi:10.7717/peerj.2874>
  to estimate the natural and fishing mortality of both species.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/klexdatr}
  \item Report bugs at \url{https://github.com/poissonconsulting/klexdatr/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID}) [data contributor]

Authors:
\itemize{
  \item Greg Andrusak [data contributor]
  \item Gary Pavan [data contributor]
  \item Sarah Stephenson [data contributor]
  \item Matt Neufeld [data contributor]
  \item Jeff Burrows [data contributor]
  \item Kerry Reed [data contributor]
  \item Robyn Irvine [data contributor]
  \item Harvey Andrusak [data contributor]
}

Other contributors:
\itemize{
  \item James Baxter [data contributor]
  \item Rob Bison [data contributor]
  \item Mike Ramsay [data contributor]
  \item Habitat Conservation Trust Foundation [copyright holder, funder]
  \item Fish and Wildlife Compensation Program [copyright holder, funder]
  \item Freshwater Fish Society of BC [copyright holder, funder]
  \item Ministry of Environment [copyright holder, data contributor]
  \item Bonneville Power Administration [copyright holder, funder]
  \item Idaho Department of Fish and Game [copyright holder]
  \item Kootenai Tribe of Idaho [copyright holder]
}

}
\keyword{internal}
